Rename the armory show week app to just armory-week
finishes https://artsyproduct.atlassian.net/browse/AR-97
